two commit

+++ Entidad/Persona.cs
@@ -11,7 +11,7 @@ namespace Entidad
         public string Nombre { get; set; }
         public int Edad { get; set; }
         public string Sexo { get; set; }
-        public decimal Pulsacion { get; set; }
+        public int Pulsacion { get; set; }
         public void CalcularPulsaciones()
         {
             if (Sexo.Equals("F") || Sexo.Equals("f"))
